done basic TB_Relationship

+++ app/Http/Controllers/BlogController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\StoreBlogRequest;
 use App\Http\Requests\UpdateBlogRequest;
 use App\Models\User;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class BlogController extends Controller
 {
@@ -19,7 +20,7 @@ class BlogController extends Controller
             $keyword = request()->keyword;
             $query->where('title', "Like", "%".$keyword."%");
             $query->orWhere("description", "Like", "%".$keyword."%");
-        })->when(request()->has("title"), function($query){
+        })->when(Auth::user()->role !== "admin", fn($query) => $query->where('user_id', Auth::id()))->when(request()->has("title"), function($query){
             $sortType = request()->title ?? 'asc';
             $query->orderBy('title', $sortType);
         })->latest("id")->paginate(7)->withQueryString();
@@ -55,9 +56,8 @@ class BlogController extends Controller
      */
     public function show(Blog $blog)
     {
-        $user = User::where('id', $blog->user_id)->first();
-
-        return view("blog.show", ['blog' => $blog, 'user' => $user]);
+        // $user = User::where('id', $blog->user_id)->first();
+        return view("blog.show", ['blog' => $blog]);
     }
 
     /**
@@ -65,6 +65,8 @@ class BlogController extends Controller
      */
     public function edit(Blog $blog)
     {
+        //Gate::authorize('blog_update', $blog); //not compatible when developing apis
+        $this->authorize('update', $blog);
         return view("blog.edit", compact('blog'));
     }
 
@@ -73,15 +75,23 @@ class BlogController extends Controller
      */
     public function update(UpdateBlogRequest $request, Blog $blog)
     {
-        // $blog->title = $request->title;
-        // $blog->description = $request->description;
-        // $blog->update();
+        // //using allows method
+        // if(!Gate::allows('blog_update', $blog)){
+        //     return abort(401);
+        // }
+
+        // if(Gate::denies('blog_update', $blog)){
+        //     return abort(403, "You're not allow to modify this blog!");
+        // }
+
+        //Gate::authorize('blog_update', $blog); //not compatible when developing apis
+
+        $this->authorize('update', $blog); // using policies
         $blog->update([
             'title' => $request->title,
             'description' => $request->description,
             'category_id' => $request->category,
         ]);
-
         return redirect()->route('blog.index')->with(['message' => "Your blog has been updated successfully!"] );
     }
 
@@ -90,6 +100,8 @@ class BlogController extends Controller
      */
     public function destroy(Blog $blog)
     {
+        //Gate::authorize('blog_delete', $blog);
+        $this->authorize('delete', $blog); // using policies
         $blog->delete();
         return redirect()->back()->with(['message' => "Your blog has been deleted successfully!"]);
     }
