+++ app/Http/Controllers/CategoryController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryController extends Controller
 {
@@ -14,6 +15,7 @@ class CategoryController extends Controller
      */
     public function index()
     {
+        // $this->authorize('viewAny', Category::class);
         $categories = Category::latest('id')->paginate(7);
         return view('category.index', compact('categories'));
 
@@ -24,6 +26,7 @@ class CategoryController extends Controller
      */
     public function create()
     {
+        // $this->authorize('create', Category::class);
         return view('category.create');
     }
 
@@ -32,6 +35,7 @@ class CategoryController extends Controller
      */
     public function store(StoreCategoryRequest $request)
     {
+        // $this->authorize('create', Category::class);
         Category::create([
             'title' => $request->title,
             'user_id' => Auth::id(),
@@ -53,6 +57,9 @@ class CategoryController extends Controller
      */
     public function edit(Category $category)
     {
+        // Gate::authorize('update', $category);
+        $this->authorize('update', Category::class);
+        // $this->authorize('cat_update', $category);
         return view('category.edit', compact('category'));
 
     }
@@ -62,6 +69,16 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+        // Gate::authorize('update', $category);
+
+        //new syntax for using policies ability
+        // if($request->user()->cannot('update', $category)){
+        //     return abort(403, "Sorry you're not allowed!");
+        // }
+
+        //another way
+        $this->authorize('update',$category);
+        // $this->authorize('cat_update', $category);
         $category->update([
             'title' => $request->title,
         ]);
@@ -75,6 +92,8 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        $this->authorize('delete', $category);
+        // $this->authorize('cat_delete', $category);
         $category->delete();
         return redirect()->back();
